Move notifications tests into their own file
